+++ csv/parse_stream_test.ts
@@ -82,12 +82,9 @@ Deno.test({
       },
       {
         name: "Separator is undefined",
-        input: "a;b;c\n",
+        input: "a,b,c\n",
+        output: [["a", "b", "c"]],
         separator: undefined,
-        error: {
-          klass: TypeError,
-          msg: "Cannot parse record: separator is required",
-        },
       },
       {
         name: "MultiLine",
@@ -418,6 +415,7 @@ x,,,
       await t.step(testCase.name, async () => {
         const options: CsvParseStreamOptions = {};
         if ("separator" in testCase) {
+          // @ts-expect-error: explicitly giving undefined
           options.separator = testCase.separator;
         }
         if ("comment" in testCase) {
@@ -510,7 +508,7 @@ Deno.test({
 
     // skipFirstRow option
     {
-      const { readable } = new CsvParseStream({ skipFirstRow: undefined });
+      const { readable } = new CsvParseStream({});
       type _ = AssertTrue<IsExact<typeof readable, ReadableStream<string[]>>>;
     }
     {
@@ -529,7 +527,7 @@ Deno.test({
 
     // columns option
     {
-      const { readable } = new CsvParseStream({ columns: undefined });
+      const { readable } = new CsvParseStream({});
       type _ = AssertTrue<IsExact<typeof readable, ReadableStream<string[]>>>;
     }
     {
@@ -550,17 +548,11 @@ Deno.test({
 
     // skipFirstRow option + columns option
     {
-      const { readable } = new CsvParseStream({
-        skipFirstRow: false,
-        columns: undefined,
-      });
+      const { readable } = new CsvParseStream({ skipFirstRow: false });
       type _ = AssertTrue<IsExact<typeof readable, ReadableStream<string[]>>>;
     }
     {
-      const { readable } = new CsvParseStream({
-        skipFirstRow: true,
-        columns: undefined,
-      });
+      const { readable } = new CsvParseStream({ skipFirstRow: true });
       type _ = AssertTrue<
         IsExact<
           typeof readable,