+++ path/_common/format.ts
@@ -9,7 +9,7 @@ export function _format(
 ): string {
   const dir: string | undefined = pathObject.dir || pathObject.root;
   const base: string = pathObject.base ||
-    (pathObject.name || "") + (pathObject.ext || "");
+    (pathObject.name ?? "") + (pathObject.ext ?? "");
   if (!dir) return base;
   if (base === sep) return dir;
   if (dir === pathObject.root) return dir + base;